% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otc_historical_contracts.R
\name{otc_historical_contracts}
\alias{otc_historical_contracts}
\title{Scrape Historical Contracts By Position}
\usage{
otc_historical_contracts(
  position = c("QB", "RB", "FB", "WR", "TE", "LT", "LG", "C", "RG", "RT", "IDL", "ED",
    "LB", "CB", "S", "K", "P", "LS")
)
}
\arguments{
\item{position}{A valid character string naming the position to scrape
historical contracts for.}
}
\value{
A tibble
}
\description{
Scrape Historical Contracts By Position
}
\examples{
\donttest{
otc_historical_contracts("QB")
}
}
