% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otc_historical_contracts.R
\name{otc_historical_contracts_all}
\alias{otc_historical_contracts_all}
\title{Scrape Historical Contracts for Multiple Positions}
\usage{
otc_historical_contracts_all(positions = NULL)
}
\arguments{
\item{positions}{A character vector with valid position names forwarded to
\code{\link[=otc_historical_contracts]{otc_historical_contracts()}}.}
}
\value{
A tibble
}
\description{
This is a wrapper around \code{\link[=otc_historical_contracts]{otc_historical_contracts()}} that
scrapes and binds multiple positions.
}
\examples{
\donttest{
# otc_historical_contracts_all()
}
}
