% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otc_player_details.R
\name{otc_player_details}
\alias{otc_player_details}
\title{Scrape Player Details}
\usage{
otc_player_details(player_url)
}
\arguments{
\item{player_url}{A valid OverTheCap player url}
}
\value{
A tibble containing draft details and contract season history
}
\description{
Scrape Player Details
}
\examples{
\donttest{
  otc_player_details("https://overthecap.com/player/aaron-rodgers/1085/")
}
}
