% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otc_player_ids.R
\name{otc_player_ids}
\alias{otc_player_ids}
\title{Retrieve player ID mappings}
\usage{
otc_player_ids(
  endpoint = Sys.getenv("OTC_PLAYERID_ENDPOINT"),
  api_key = Sys.getenv("OTC_API_KEY")
)
}
\arguments{
\item{endpoint}{endpoint (defaults to environment variable OTC_PLAYERID_ENDPOINT)}

\item{api_key}{api key (defaults to environment variable OTC_API_KEY)}
}
\description{
Retrieve player ID mappings
}
