% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otc_team_contracts.R
\name{otc_team_contracts}
\alias{otc_team_contracts}
\title{Scrape Historical Contracts By Position}
\usage{
otc_team_contracts(team)
}
\arguments{
\item{team}{A valid character string naming the position to scrape
historical contracts for.}
}
\value{
A tibble
}
\description{
Scrape Historical Contracts By Position
}
\examples{
\donttest{
otc_historical_contracts("QB")
}
}
