% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotc-package.R
\docType{package}
\name{rotc-package}
\alias{rotc}
\alias{rotc-package}
\title{rotc: Functions to Access OTC Data}
\description{
A set of functions to access over the cap data.
}
\author{
\strong{Maintainer}: Carl Sebastian \email{mrcaseb@gmail.com}

}
\keyword{internal}
